Move AsyncQueryLabel to th frontend as it is causing issues with framework publishing
